+++ lib/factorix/cli.rb
@@ -17,6 +17,7 @@ module Factorix
     register "man", Commands::Man
     register "launch", Commands::Launch
     register "path", Commands::Path
+    register "download", Commands::Download
     register "completion", Commands::Completion
     register "mod check", Commands::MOD::Check
     register "mod list", Commands::MOD::List